% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raw_lst_data.R
\name{build_url}
\alias{build_url}
\title{Build the url for the migration data from the statisik}
\usage{
build_url(gkz, sex, nationality, year_min, year_max)
}
\description{
Build the url for the migration data from the statisik
}
