% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpi_data_raw.R
\docType{data}
\name{data_vpi_raw}
\alias{data_vpi_raw}
\title{Data on the yearly VPI of Austria}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 59 rows and 14 columns.
}
\source{
\url{https://www.wko.at/zahlen-daten-fakten/verbraucherpreisindex#heading_Verbraucherpreisindex__VPI____Indexentwicklung}
}
\usage{
data_vpi_raw
}
\description{
Data on the yearly VPI of Austria
}
\keyword{datasets}
