% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_gemeinde_single.R
\name{get_clean_gemeinde}
\alias{get_clean_gemeinde}
\title{Get the history of a current gemeinde}
\usage{
get_clean_gemeinde(gem, years_range, data_all)
}
\arguments{
\item{data_all}{}
}
\description{
Get the history of a current gemeinde
}
