% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_gemeinden.R
\name{get_clean_gemeinden_raw}
\alias{get_clean_gemeinden_raw}
\title{Create one File that tracks the changes of the gemeinden}
\usage{
get_clean_gemeinden_raw(
  year_gemeinde_data = 2021,
  path_to_aenderungen = NULL,
  out_dir = NULL,
  TEST = FALSE
)
}
\arguments{
\item{path_to_aenderungen}{}
}
\description{
Create one File that tracks the changes of the gemeinden
}
