% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historic_income.R
\name{get_historic_income}
\alias{get_historic_income}
\title{Get the income for Gemeinden with Änderungen}
\usage{
get_historic_income(raw_data, historic_gemeindestaende)
}
\arguments{
\item{raw_data}{Raw income data for \strong{one variable} (e.g. the Median)}

\item{gems_changes}{a list of changes for each gemeinde}
}
\description{
Get the income for Gemeinden with Änderungen
}
