% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_data_statistik.R
\name{read_lst_data}
\alias{read_lst_data}
\title{Process the statistik lst data}
\usage{
read_lst_data(
  dir_with_excels = NULL,
  sex = "Zusammen",
  type = "Nettobezüge",
  variable = "Median",
  out_dir = NULL
)
}
\arguments{
\item{dir_with_excels}{directory with all the xlsx files}

\item{sex}{choose between (Frauen, Männer, Zusammen (default))}

\item{type}{choose between (Bruttobezüge, Nettobezüge (default), SV-Beiträge, einbehaltene Lohnsteuer)}

\item{variable}{Anzahl, Median (default), Mittelwert, P10, P20, P30, P40, P60, P70, P80, P90, Summe}
}
\description{
Loads, for each year, the Lohnsteuerstatistikdaten from the Statistik Austria
for a given sex, type of income (netto, brutto) and statistic value (e.g. median)
}
